+++ hotdrop_direct/driver-examples.spec.js
@@ -1,6 +1,6 @@
 const driver = require("./index.js");
 const examples = require("./examples.json");
-const es5dot1driver = require("./index-es5dot1.js");
+const es5dot1driver = require("./index-es5.js");
 
 /*..............
 Test suites
